scripts/update-comment/api_comment_updater/requirements.txt
# HTML解析
beautifulsoup4==4.12.2
lxml==4.9.3

# 配置管理
pyyaml==6.0.1
jsonschema==4.19.0

# 命令行界面
click==8.1.7
rich==13.5.2

# 日志处理
loguru==0.7.2

# 测试和质量
pytest==7.4.2
black==23.7.0
flake8==6.0.0
